Added new permission validations to proposals module. Fixed some minor errors in styling,

+++ src/apps/ecidadania/proposals/views/proposalsets.py
@@ -39,6 +39,7 @@ from django.utils.translation import ugettext_lazy as _
 from django.utils.decorators import method_decorator
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render_to_response, get_object_or_404, redirect
+from django.core.exceptions import PermissionDenied
 
 from apps.ecidadania.proposals import url_names as urln_prop
 from core.spaces import url_names as urln_space
@@ -63,6 +64,14 @@ class AddProposalInSet(FormView):
     form_class = ProposalFormInSet
     template_name = 'proposals/proposal_form_in_set.html'
 
+    def dispatch(self, request, *args, **kwargs):
+        space = get_object_or_404(Space, url=kwargs['space_url'])
+
+        if request.user.has_perm('view_space', space):
+            return super(AddProposalInSet, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied
+
     def get_success_url(self):
         space = self.kwargs['space_url']
         return reverse(urln_space.SPACE_INDEX, kwargs={'space_url': space})
@@ -83,10 +92,7 @@ class AddProposalInSet(FormView):
         self.field = ProposalField.objects.filter(proposalset=self.kwargs['set_id'])
         context['get_place'] = self.space
         context['form_field'] = [f_name.field_name for f_name in self.field]
-        return context
-
-    def dispatch(self, *args, **kwargs):
-        return super(AddProposalInSet, self).dispatch(*args, **kwargs)
+        return contextç
 
 
 def add_proposal_field(request, space_url):
@@ -102,23 +108,28 @@ def add_proposal_field(request, space_url):
     :context:form, get_place, prop_fields, form_data, prop_fields
 
     """
-    form = ProposalFieldForm(request.POST or None)
     get_place = get_object_or_404(Space, url=space_url)
-    if request.method == 'POST':
-        if form.is_valid():
-            form_data = form.save()
-            proposal_fields = ProposalField.objects.filter(
-                proposalset=form_data.proposalset)
-            return render_to_response("proposals/proposal_add_fields.html",
-                {'form_data': form_data,
-                 'get_place': get_place,
-                 'prop_fields': proposal_fields,
-                 'form': form},
-                context_instance=RequestContext(request))
 
-    return render_to_response("proposals/proposal_add_fields.html",
-        {'form': form, 'get_place': get_place},
-        context_instance=RequestContext(request))
+    if (request.user.has_perm('admin_space', get_place) or
+        request.user.has_perm('mod_space', get_place)):
+        form = ProposalFieldForm(request.POST or None)
+        if request.method == 'POST':
+            if form.is_valid():
+                form_data = form.save()
+                proposal_fields = ProposalField.objects.filter(
+                    proposalset=form_data.proposalset)
+                return render_to_response("proposals/proposal_add_fields.html",
+                    {'form_data': form_data,
+                     'get_place': get_place,
+                     'prop_fields': proposal_fields,
+                     'form': form},
+                    context_instance=RequestContext(request))
+
+        return render_to_response("proposals/proposal_add_fields.html",
+            {'form': form, 'get_place': get_place},
+            context_instance=RequestContext(request))
+    else:
+        raise PermissionDenied
 
 
 def delete_proposal_field(request, space_url):
@@ -132,24 +143,29 @@ def delete_proposal_field(request, space_url):
     :context: d_form, get_place, delete_field
 
     """
-    d_form = ProposalFieldDeleteForm(request.POST or None)
     get_place = get_object_or_404(Space, url=space_url)
-    if request.method == 'POST':
-        if d_form.is_valid():
-            form_data = d_form.save(commit=False)
-            delete_field = ProposalField.objects.filter(
-                proposalset=form_data.proposalset,
-                field_name=form_data.field_name)
-            delete_field.delete()
-            return render_to_response(
-                "proposals/proposalform_remove_field.html",
-                {'form': d_form, 'get_place': get_place,
-                 'deleted_field': form_data},
-                context_instance=RequestContext(request))
 
-    return render_to_response("proposals/proposalform_remove_field.html",
-        {'form': d_form, 'get_place': get_place},
-        context_instance=RequestContext(request))
+    if (request.user.has_perm('admin_space', get_place) or
+        request.user.has_perm('mod_space', get_place)):
+        d_form = ProposalFieldDeleteForm(request.POST or None)
+        if request.method == 'POST':
+            if d_form.is_valid():
+                form_data = d_form.save(commit=False)
+                delete_field = ProposalField.objects.filter(
+                    proposalset=form_data.proposalset,
+                    field_name=form_data.field_name)
+                delete_field.delete()
+                return render_to_response(
+                    "proposals/proposalform_remove_field.html",
+                    {'form': d_form, 'get_place': get_place,
+                     'deleted_field': form_data},
+                    context_instance=RequestContext(request))
+
+        return render_to_response("proposals/proposalform_remove_field.html",
+            {'form': d_form, 'get_place': get_place},
+            context_instance=RequestContext(request))
+    else:
+        raise PermissionDenied
 
 
 def proposal_to_set(request, space_url):
@@ -164,21 +180,26 @@ def proposal_to_set(request, space_url):
 
     """
     get_place = get_object_or_404(Space, url=space_url)
-    sel_form = ProposalSetSelectForm(request.POST or None)
-    # We change here the queryset, so only the proposalsets of this space
-    # appear on the list.
-    sel_form.fields['proposalset'].queryset = ProposalSet.objects.filter(
-        space=get_place)
 
-    if request.method == 'POST':
-        if sel_form.is_valid():
-            pset = request.POST['proposalset']
-            return HttpResponseRedirect(reverse(urln_prop.PROPOSAL_ADD_INSET,
-                kwargs={'space_url': space_url, 'set_id': pset}))
+    if (request.user.has_perm('admin_space', get_place) or
+        request.user.has_perm('mod_space', get_place)):
+        sel_form = ProposalSetSelectForm(request.POST or None)
+        # We change here the queryset, so only the proposalsets of this space
+        # appear on the list.
+        sel_form.fields['proposalset'].queryset = ProposalSet.objects.filter(
+            space=get_place)
 
-    return render_to_response("proposals/proposalset_select_form.html",
-        {'form': sel_form, 'get_place': get_place},
-        context_instance=RequestContext(request))
+        if request.method == 'POST':
+            if sel_form.is_valid():
+                pset = request.POST['proposalset']
+                return HttpResponseRedirect(reverse(urln_prop.PROPOSAL_ADD_INSET,
+                    kwargs={'space_url': space_url, 'set_id': pset}))
+
+        return render_to_response("proposals/proposalset_select_form.html",
+            {'form': sel_form, 'get_place': get_place},
+            context_instance=RequestContext(request))
+    else:
+        raise PermissionDenied
 
 
 def mergedproposal_to_set(request, space_url):
@@ -190,18 +211,22 @@ def mergedproposal_to_set(request, space_url):
     :context:form, get_place
 
     """
-
-    sel_form = ProposalSetSelectForm(request.POST or None)
     get_place = get_object_or_404(Space, url=space_url)
 
-    if request.method == 'POST':
-        if sel_form.is_valid():
-            pset = request.POST['proposalset']
-            return HttpResponseRedirect(reverse(urln_prop.PROPOSAL_MERGED, kwargs={'space_url': space_url, 'set_id': pset}))
+    if (request.user.has_perm('admin_space', get_place) or
+        request.user.has_perm('mod_space', get_place)):
+        sel_form = ProposalSetSelectForm(request.POST or None)
 
-    return render_to_response("proposals/mergedproposal_in_set.html",
-        {'form': sel_form, 'get_place': get_place},
-        context_instance=RequestContext(request))
+        if request.method == 'POST':
+            if sel_form.is_valid():
+                pset = request.POST['proposalset']
+                return HttpResponseRedirect(reverse(urln_prop.PROPOSAL_MERGED, kwargs={'space_url': space_url, 'set_id': pset}))
+
+        return render_to_response("proposals/mergedproposal_in_set.html",
+            {'form': sel_form, 'get_place': get_place},
+            context_instance=RequestContext(request))
+    else:
+        raise PermissionDenied
 
 
 #
@@ -220,6 +245,14 @@ class ListProposalSet(ListView):
     """
     paginate_by = 20
     context_object_name = 'setlist'
+
+    def dispatch(self, request, *args, **kwargs):
+        space = get_object_or_404(Space, url=kwargs['space_url'])
+
+        if request.user.has_perm('view_space', space):
+            return super(ListProposalSet, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied
 
     def get_queryset(self):
         cur_space = self.kwargs['space_url']
@@ -246,6 +279,14 @@ class ViewProposalSet(ListView):
     paginate_by = 50
     context_object_name = 'proposalset'
     template_name = 'proposals/proposalset_detail.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        space = get_object_or_404(Space, url=kwargs['space_url'])
+
+        if request.user.has_perm('view_space', space):
+            return super(ViewProposalSet, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied
 
     def get_queryset(self):
         place = get_object_or_404(Space, url=self.kwargs['space_url'])
@@ -275,6 +316,15 @@ class AddProposalSet(FormView):
     form_class = ProposalSetForm
     template_name = 'proposals/proposalset_form.html'
 
+    def dispatch(self, request, *args, **kwargs):
+        space = get_object_or_404(Space, url=kwargs['space_url'])
+
+        if (request.user.has_perm('admin_space', space) or
+            request.user.has_perm('mod_space', space)):
+            return super(AddProposalSet, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied
+
     def get_success_url(self):
         space = self.kwargs['space_url']
         return reverse(urln_space.SPACE_INDEX, kwargs={'space_url': space})
@@ -298,10 +348,6 @@ class AddProposalSet(FormView):
         context['get_place'] = self.space
         return context
 
-    @method_decorator(permission_required('proposals.add_proposalset'))
-    def dispatch(self, *args, **kwargs):
-        return super(AddProposalSet, self).dispatch(*args, **kwargs)
-
 
 class EditProposalSet(UpdateView):
 
@@ -315,6 +361,15 @@ class EditProposalSet(UpdateView):
     """
     model = ProposalSet
     template_name = 'proposals/proposalset_form.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        space = get_object_or_404(Space, url=kwargs['space_url'])
+
+        if (request.user.has_perm('admin_space', space) or
+            request.user.has_perm('mod_space', space)):
+            return super(EditProposalSet, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied
 
     def get_success_url(self):
         space = self.kwargs['space_url']
@@ -331,10 +386,6 @@ class EditProposalSet(UpdateView):
         context['get_place'] = get_object_or_404(Space, url=self.kwargs['space_url'])
         return context
 
-    @method_decorator(permission_required('proposals.change_proposalset'))
-    def dispatch(self, *args, **kwargs):
-        return super(EditProposalSet, self).dispatch(*args, **kwargs)
-
 
 class DeleteProposalSet(DeleteView):
 
@@ -346,6 +397,15 @@ class DeleteProposalSet(DeleteView):
     :rtype: Confirmation
     :context: get_place
     """
+    def dispatch(self, request, *args, **kwargs):
+        space = get_object_or_404(Space, url=kwargs['space_url'])
+
+        if (request.user.has_perm('admin_space', space) or
+            request.user.has_perm('mod_space', space)):
+            return super(DeleteProposalSet, self).dispatch(request, *args, **kwargs)
+        else:
+            raise PermissionDenied
+
     def get_object(self):
         return get_object_or_404(ProposalSet, pk=self.kwargs['set_id'])
